+++ CotEditor/Sources/EditorTextView.swift
@@ -121,8 +121,10 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         super.init(coder: coder)
         
         // workaround for: the text selection highlight can remain between lines (2017-09 macOS 10.13).
-        self.scaleUnitSquare(to: NSSize(width: 0.5, height: 0.5))
-        self.scaleUnitSquare(to: self.convert(.unit, from: nil))  // reset scale
+        if NSAppKitVersion.current <= .macOS10_13 {
+            self.scaleUnitSquare(to: NSSize(width: 0.5, height: 0.5))
+            self.scaleUnitSquare(to: self.convert(.unit, from: nil))  // reset scale
+        }
         
         // setup layoutManager and textContainer
         let layoutManager = LayoutManager()
@@ -300,7 +302,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         self.needsUpdateLineHighlight = true
         
         // retry completion if needed
-        //   -> Flag is set in `insertCompletion:forPartialWordRange:movement:isFinal:`
+        //   -> Flag is set in `insertCompletion(_:forPartialWordRange:movement:isFinal:)`
         if self.needsRecompletion {
             self.needsRecompletion = false
             self.completionTask.schedule(delay: .milliseconds(50))
@@ -311,73 +313,68 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     /// on inputting text (NSTextInputClient Protocol)
     override func insertText(_ string: Any, replacementRange: NSRange) {
         
-        // do not use this method for programmatical insertion.
+        // do not use this method for programmatic insertion.
         
-        // cast NSAttributedString to String in order to make sure input string is plain-text
-        guard let plainString: String = {
-            switch string {
-            case let attrString as NSAttributedString:
-                return attrString.string
-            case let string as String:
-                return string
-            default: return nil
+        // sanitize input to plain string
+        let plainString: String = {
+            // cast input to String
+            let input: String = {
+                switch string {
+                case let attrString as NSAttributedString:
+                    return attrString.string
+                case let string as String:
+                    return string
+                default: preconditionFailure()
+                }
+            }()
+            
+            // swap '¥' with '\' if needed
+            if UserDefaults.standard[.swapYenAndBackSlash] {
+                switch input {
+                case "\\": return "¥"
+                case "¥": return "\\"
+                default: break
+                }
             }
-            }() else { return super.insertText(string, replacementRange: replacementRange) }
-        
-        // swap '¥' with '\' if needed
-        if UserDefaults.standard[.swapYenAndBackSlash], plainString.count == 1 {
-            switch plainString {
-            case "\\":
-                return super.insertText("¥", replacementRange: replacementRange)
-            case "¥":
-                return super.insertText("\\", replacementRange: replacementRange)
-            default: break
-            }
-        }
+            
+            return input
+        }()
         
         // balance brackets and quotes
-        if self.balancesBrackets && replacementRange.length == 0,
-            plainString.unicodeScalars.count == 1,
-            let firstChar = plainString.first,
-            let pair = self.matchingBracketPairs.first(where: { $0.begin == firstChar })
-        {
-            // wrap selection with brackets if some text is selected
-            if self.selectedRange.length > 0 {
-                self.surroundSelections(begin: String(pair.begin), end: String(pair.end))
-                return
+        if self.balancesBrackets, replacementRange.length == 0 {
+            // with opening symbol input
+            if let pair = self.matchingBracketPairs.first(where: { String($0.begin) == plainString }) {
+                // wrap selection with brackets if some text is selected
+                if self.rangeForUserTextChange.length > 0 {
+                    self.surroundSelections(begin: String(pair.begin), end: String(pair.end))
+                    return
+                }
                 
-            // check if insertion point is in a word
-            } else if
-                !CharacterSet.alphanumerics.contains(self.characterAfterInsertion ?? UnicodeScalar(0)),
-                !(pair.begin == pair.end && CharacterSet.alphanumerics.contains(self.characterBeforeInsertion ?? UnicodeScalar(0)))  // for "
-            {
-                let pairedBrackets = String(pair.begin) + String(pair.end)
+                // insert bracket pair if insertion point is not in a word
+                if !CharacterSet.alphanumerics.contains(self.characterAfterInsertion ?? UnicodeScalar(0)),
+                    !(pair.begin == pair.end && CharacterSet.alphanumerics.contains(self.characterBeforeInsertion ?? UnicodeScalar(0)))  // for "
+                {
+                    super.insertText(String(pair.begin) + String(pair.end), replacementRange: replacementRange)
+                    self.setSelectedRangesWithUndo([NSRange(location: self.selectedRange.location - 1, length: 0)])
+                    self.textStorage?.addAttribute(.autoBalancedClosingBracket, value: true,
+                                                   range: NSRange(location: self.selectedRange.location, length: 1))
+                    return
+                }
+            }
             
-                super.insertText(pairedBrackets, replacementRange: replacementRange)
-                self.selectedRange = NSRange(location: self.selectedRange.location - 1, length: 0)
-                
-                // set flag
-                self.textStorage?.addAttribute(.autoBalancedClosingBracket, value: true,
-                                               range: NSRange(location: self.selectedRange.location, length: 1))
-                
+            // just move cursor if closing bracket is already typed
+            if BracePair.braces.contains(where: { String($0.end) == plainString }),  // ignore "
+                plainString.unicodeScalars.first == self.characterAfterInsertion,
+                self.textStorage?.attribute(.autoBalancedClosingBracket, at: self.selectedRange.location, effectiveRange: nil) as? Bool ?? false
+            {
+                self.selectedRange.location += 1
                 return
             }
-        }
-        
-        // just move cursor if closed bracket is already typed
-        if self.balancesBrackets && replacementRange.length == 0,
-            let nextCharacter = self.characterAfterInsertion,
-            let firstCharacter = plainString.first, firstCharacter == Character(nextCharacter),
-            BracePair.braces.contains(where: { $0.end == firstCharacter }),  // ignore "
-            self.textStorage?.attribute(.autoBalancedClosingBracket, at: self.selectedRange.location, effectiveRange: nil) as? Bool ?? false
-        {
-            self.selectedRange.location += 1
-            return
         }
         
         // smart outdent with '}'
-        if self.isAutomaticIndentEnabled, self.isSmartIndentEnabled,
-            replacementRange.length == 0, plainString == "}",
+        if self.isAutomaticIndentEnabled, self.isSmartIndentEnabled, replacementRange.length == 0,
+            plainString == "}",
             let insertionIndex = Range(self.selectedRange, in: self.string)?.upperBound
         {
             let lineRange = self.string.lineRange(at: insertionIndex)
@@ -411,17 +408,15 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     override func insertTab(_ sender: Any?) {
         
         // indent with tab key
-        if UserDefaults.standard[.indentWithTabKey], self.selectedRange.length > 0 {
+        if UserDefaults.standard[.indentWithTabKey], self.rangeForUserTextChange.length > 0 {
             self.indent()
             return
         }
         
         if self.isAutomaticTabExpansionEnabled {
-            let column = self.string.column(of: self.rangeForUserTextChange.location, tabWidth: self.tabWidth)
-            let length = self.tabWidth - (column % self.tabWidth)
-            let spaces = String(repeating: " ", count: length)
+            let softTab = self.string.softTab(at: self.rangeForUserTextChange.location, tabWidth: self.tabWidth)
             
-            return super.insertText(spaces, replacementRange: self.rangeForUserTextChange)
+            return super.insertText(softTab, replacementRange: self.rangeForUserTextChange)
         }
         
         super.insertTab(sender)
@@ -444,36 +439,37 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     /// insert new line & perform auto-indent
     override func insertNewline(_ sender: Any?) {
         
-        guard self.isAutomaticIndentEnabled else {
-            return super.insertNewline(sender)
-        }
+        guard
+            self.isEditable,
+            self.isAutomaticIndentEnabled
+            else { return super.insertNewline(sender) }
         
-        let indentRange = self.string.rangeOfIndent(at: self.selectedRange.location)
+        let indentRange = self.string.rangeOfIndent(at: self.rangeForUserTextChange.location)
         
         // don't auto-indent if indent is selected (2008-12-13)
-        guard indentRange.length == 0 || indentRange != self.selectedRange else {
+        guard indentRange.length == 0 || indentRange != self.rangeForUserTextChange else {
             return super.insertNewline(sender)
         }
         
         let indent: String = {
-            guard let baseIndentRange = indentRange.intersection(NSRange(0..<self.selectedRange.location)) else {
+            guard let autoIndentRange = indentRange.intersection(NSRange(0..<self.rangeForUserTextChange.location)) else {
                 return ""
             }
-            return (self.string as NSString).substring(with: baseIndentRange)
+            return (self.string as NSString).substring(with: autoIndentRange)
         }()
         
-        // calculation for smart indent
-        var shouldIncreaseIndentLevel = false
-        var shouldExpandBlock = false
+        // check if smart indent required
+        let shouldExpandBlock: Bool
+        let shouldIncreaseIndentLevel: Bool
         if self.isSmartIndentEnabled {
             let lastCharacter = self.characterBeforeInsertion
             let nextCharacter = self.characterAfterInsertion
             
-            // expand idnent block if returned inside `{}`
             shouldExpandBlock = (lastCharacter == "{" && nextCharacter == "}")
-            
-            // increace font indent level if the character just before the return is `:` or `{`
             shouldIncreaseIndentLevel = (lastCharacter == ":" || lastCharacter == "{")
+        } else {
+            shouldExpandBlock = false
+            shouldIncreaseIndentLevel = false
         }
         
         super.insertNewline(sender)
@@ -485,13 +481,12 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         
         // smart indent
         if shouldExpandBlock {
-            self.insertTab(sender)
-            let selection = self.selectedRange
+            let selectedRanges = self.selectedRanges
             super.insertNewline(sender)
             super.insertText(indent, replacementRange: self.rangeForUserTextChange)
-            self.selectedRange = selection
-            
-        } else if shouldIncreaseIndentLevel {
+            self.selectedRanges = selectedRanges
+        }
+        if shouldIncreaseIndentLevel {
             self.insertTab(sender)
         }
     }
@@ -500,36 +495,28 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     /// delete & adjust indent
     override func deleteBackward(_ sender: Any?) {
         
-        defer {
-            super.deleteBackward(sender)
-        }
-        
-        guard self.selectedRange.length == 0 else { return }
-        
-        let location = self.selectedRange.location
+        guard self.isEditable else { return super.deleteBackward(sender) }
         
         // delete tab
         if self.isAutomaticTabExpansionEnabled,
-            self.string.rangeOfIndent(at: location).upperBound >= location
+            let deletionRange = self.string.rangeForSoftTabDeletion(in: self.rangeForUserTextChange, tabWidth: self.tabWidth)
         {
-            let column = self.string.column(of: location, tabWidth: self.tabWidth)
-            let targetLength = self.tabWidth - (column % self.tabWidth)
-            let targetRange = NSRange((location - targetLength)..<location)
-            
-            if location >= targetLength,
-                (self.string as NSString).substring(with: targetRange) == String(repeating: " ", count: targetLength) {
-                self.selectedRange = targetRange
-            }
+            self.setSelectedRangesWithUndo(self.selectedRanges)
+            self.selectedRange = deletionRange
         }
         
         // balance brackets
         if self.balancesBrackets,
+            self.selectedRange.length == 0,
             let lastCharacter = self.characterBeforeInsertion,
             let nextCharacter = self.characterAfterInsertion,
             self.matchingBracketPairs.contains(where: { $0.begin == Character(lastCharacter) && $0.end == Character(nextCharacter) })
         {
-            self.selectedRange = NSRange(location: location - 1, length: 2)
+            self.setSelectedRangesWithUndo(self.selectedRanges)
+            self.selectedRange = NSRange(location: self.rangeForUserTextChange.location - 1, length: 2)
         }
+        
+        super.deleteBackward(sender)
     }
     
     
@@ -678,7 +665,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         // minimize drawing area on non-opaque background
         // -> Otherwise, all textView (from the top to the bottom) is everytime drawn
         //    and it affects to the drawing performance on a large document critically. (2017-03 macOS 10.12)
-        let dirtyRect = self.drawsBackground ? dirtyRect: self.visibleRect
+        let dirtyRect = self.drawsBackground ? dirtyRect : self.visibleRect
         
         super.draw(dirtyRect)
         
@@ -719,7 +706,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
             let layoutManager = self.layoutManager,
             let textContainer = self.textContainer
         {
-            layoutManager.ensureLayout(forCharacterRange: NSRange(location: 0, length: range.upperBound))
+            layoutManager.ensureLayout(forCharacterRange: NSRange(..<range.upperBound))
             let glyphRange = layoutManager.glyphRange(forCharacterRange: range, actualCharacterRange: nil)
             let glyphRect = layoutManager.boundingRect(forGlyphRange: glyphRange, in: textContainer)
                 .offset(by: self.textContainerOrigin)
@@ -972,7 +959,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     @objc var tabWidth: Int {
         
         didSet {
-            if tabWidth == 0 {
+            if tabWidth <= 0 {
                 tabWidth = oldValue
             }
             guard tabWidth != oldValue else { return }
@@ -987,7 +974,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     var lineHeight: CGFloat {
         
         didSet {
-            if lineHeight == 0 {
+            if lineHeight <= 0 {
                 lineHeight = oldValue
             }
             guard lineHeight != oldValue else { return }
@@ -1124,14 +1111,14 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
     /// input an Yen sign (¥)
     @IBAction func inputYenMark(_ sender: Any?) {
         
-        super.insertText("¥", replacementRange: self.rangeForUserTextChange)
+        super.insertText("¥", replacementRange: .notFound)
     }
     
     
     ///input a backslash (\\)
     @IBAction func inputBackSlash(_ sender: Any?) {
         
-        super.insertText("\\", replacementRange: self.rangeForUserTextChange)
+        super.insertText("\\", replacementRange: .notFound)
     }
     
     
@@ -1317,12 +1304,19 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         
         // -> use own dataDetector instead of `checkTextInDocument(_:)` due to performance issue (2018-07)
         let detector = try! NSDataDetector(types: NSTextCheckingResult.CheckingType.link.rawValue)
+        let range = self.string.nsRange
         
-        detector.enumerateMatches(in: self.string, range: self.string.nsRange) { (result, _, _) in
+        self.textStorage?.removeAttribute(.link, range: range)
+        
+        detector.enumerateMatches(in: self.string, range: range) { (result, _, _) in
             guard let result = result, let url = result.url else { return }
             
             self.textStorage?.addAttribute(.link, value: url, range: result.range)
         }
+        
+        // ensure layout to avoid unwanted scroll with cursor move after pasting something
+        // at the latter part of the document. (2018-10 macOS 10.14)
+        self.layoutManager?.ensureLayout(forCharacterRange: range)
     }
     
     
@@ -1403,15 +1397,15 @@ extension EditorTextView {
         
         let range = super.rangeForUserCompletion
         
-        guard !self.syntaxCompletionWords.isEmpty else { return range }
-        
         let firstLetters = self.syntaxCompletionWords.compactMap { $0.unicodeScalars.first }
+        let firstLetterSet = CharacterSet(firstLetters)
         
-        // expand range until hitting to a character that isn't in the word completion candidates
+        // expand range until hitting a character that isn't in the word completion candidates
         guard
+            !firstLetterSet.isEmpty,
             !self.string.isEmpty,
             let characterRange = Range(range, in: self.string),
-            let index = self.string.rangeOfCharacter(from: CharacterSet(firstLetters).inverted, options: .backwards, range: self.string.startIndex..<characterRange.upperBound)?.upperBound
+            let index = self.string[..<characterRange.upperBound].rangeOfCharacter(from: firstLetterSet.inverted, options: .backwards)?.upperBound
             else { return range }
         
         return NSRange(index..<characterRange.upperBound, in: self.string)